Pokemon/requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
gunicorn==20.1.0
flask_login == 0.6.3
requests == 2.31.0
psycopg2-binary == 2.9.9

